Change all *int8_t to *int16_t since pqxx doesn't support it

+++ src/messages/message.cpp
@@ -90,7 +90,7 @@ tuple<uint32_t, unique_ptr<message<UseJsonAsReturnType> const>> message<UseJson>
         }
         case login_response_message<UseJson>::id:
         {
-            int8_t admin_status;
+            int16_t admin_status;
             uint64_t user_id;
             archive(cereal::make_nvp("admin_status", admin_status), cereal::make_nvp("user_id", user_id));
             return make_tuple(message_id, make_unique<login_response_message<UseJsonAsReturnType>>(sender, admin_status, user_id));
@@ -106,7 +106,7 @@ tuple<uint32_t, unique_ptr<message<UseJsonAsReturnType> const>> message<UseJson>
         }
         case register_response_message<UseJson>::id:
         {
-            int8_t admin_status;
+            int16_t admin_status;
             uint64_t user_id;
             archive(cereal::make_nvp("admin_status", admin_status), cereal::make_nvp("user_id", user_id));
             return make_tuple(message_id, make_unique<register_response_message<UseJsonAsReturnType>>(sender, admin_status, user_id));
