+++ src/messages/user_access_control/register_response_message.cpp
@@ -27,7 +27,7 @@ using namespace std;
 using namespace roa;
 
 template <bool UseJson>
-register_response_message<UseJson>::register_response_message(message_sender sender, int8_t admin_status, uint64_t user_id) noexcept
+register_response_message<UseJson>::register_response_message(message_sender sender, int16_t admin_status, uint64_t user_id) noexcept
         : message<UseJson>(sender), admin_status(admin_status), user_id(user_id) {}
 
 template <bool UseJson>
